chainlit/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "a professional health agemt"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai>=1.92.3",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
]
